test(contract): move config leaderboard and market suites
