+++ contract/src/market.rs
@@ -165,8 +165,7 @@ pub fn emit_market_resolved(env: &Env, market_id: u64, resolved_outcome: Symbol)
 
 /// Calculate price of outcome A in terms of outcome B.
 /// Returns price with 6 decimal precision (multiplied by 1_000_000).
-#[allow(dead_code)]
-fn calculate_price(reserve_a: i128, reserve_b: i128) -> Result<i128, InsightArenaError> {
+pub fn calculate_price(reserve_a: i128, reserve_b: i128) -> Result<i128, InsightArenaError> {
     if reserve_a <= 0 || reserve_b <= 0 {
         return Err(InsightArenaError::InvalidInput);
     }
@@ -178,6 +177,25 @@ fn calculate_price(reserve_a: i128, reserve_b: i128) -> Result<i128, InsightAren
         .ok_or(InsightArenaError::Overflow)?;
 
     Ok(price)
+}
+
+fn has_duplicate_outcomes(outcomes: &Vec<Symbol>) -> bool {
+    let mut index: u32 = 0;
+    while index < outcomes.len() {
+        let outcome = outcomes.get(index).unwrap();
+        let mut next_index = index + 1;
+
+        while next_index < outcomes.len() {
+            if outcomes.get(next_index) == Some(outcome.clone()) {
+                return true;
+            }
+            next_index += 1;
+        }
+
+        index += 1;
+    }
+
+    false
 }
 
 // ── Entry-point logic ─────────────────────────────────────────────────────────
@@ -217,6 +235,9 @@ pub fn create_market(
 
     // ── Guard 5: at least two outcomes required ───────────────────────────────
     if params.outcomes.len() < 2 {
+        return Err(InsightArenaError::InvalidInput);
+    }
+    if has_duplicate_outcomes(&params.outcomes) {
         return Err(InsightArenaError::InvalidInput);
     }
 
@@ -492,11 +513,10 @@ pub fn cancel_market(env: &Env, caller: Address, market_id: u64) -> Result<(), I
         .set(&DataKey::Market(market_id), &market);
     bump_market(env, market_id);
 
-    // ── Iterate all predictors and issue refunds ──────────────────────────────
-    let predictors: Vec<Address> = env
+    let predictors = env
         .storage()
         .persistent()
-        .get(&DataKey::PredictorList(market_id))
+        .get::<DataKey, Vec<Address>>(&DataKey::PredictorList(market_id))
         .unwrap_or_else(|| Vec::new(env));
 
     for predictor in predictors.iter() {
@@ -506,761 +526,7 @@ pub fn cancel_market(env: &Env, caller: Address, market_id: u64) -> Result<(), I
         }
     }
 
-    // ── Emit MarketCancelled event ────────────────────────────────────────────
     emit_market_cancelled(env, market_id, &caller);
 
     Ok(())
-}
-
-// ── Tests ─────────────────────────────────────────────────────────────────────
-
-#[cfg(test)]
-mod market_tests {
-    use soroban_sdk::testutils::{Address as _, Ledger as _};
-    use soroban_sdk::{symbol_short, vec, Address, Env, String, Symbol, Vec};
-
-    use crate::storage_types::DataKey;
-    use crate::{InsightArenaContract, InsightArenaContractClient, InsightArenaError};
-
-    use super::CreateMarketParams;
-
-    #[test]
-    fn test_calculate_price_equal_reserves() {
-        // Reserves: 1000/1000 -> Expected: 1_000_000
-        let price = super::calculate_price(1000, 1000).unwrap();
-        assert_eq!(price, 1_000_000);
-    }
-
-    #[test]
-    fn test_calculate_price_double() {
-        // Reserves: 1000/2000 -> Expected: 2_000_000
-        let price = super::calculate_price(1000, 2000).unwrap();
-        assert_eq!(price, 2_000_000);
-    }
-
-    #[test]
-    fn test_calculate_price_half() {
-        // Reserves: 2000/1000 -> Expected: 500_000
-        let price = super::calculate_price(2000, 1000).unwrap();
-        assert_eq!(price, 500_000);
-    }
-
-    #[test]
-    fn test_calculate_price_precision() {
-        // Reserves: 3000/1000 -> Expected: 333_333
-        let price = super::calculate_price(3000, 1000).unwrap();
-        assert_eq!(price, 333_333);
-    }
-
-    /// Register a mock XLM token (Stellar Asset Contract) and return its address.
-    fn register_token(env: &Env) -> Address {
-        let token_admin = Address::generate(env);
-        env.register_stellar_asset_contract_v2(token_admin)
-            .address()
-    }
-
-    fn deploy(env: &Env) -> InsightArenaContractClient<'_> {
-        let id = env.register(InsightArenaContract, ());
-        let client = InsightArenaContractClient::new(env, &id);
-        let admin = Address::generate(env);
-        let oracle = Address::generate(env);
-        let xlm_token = register_token(env);
-        env.mock_all_auths();
-        client.initialize(&admin, &oracle, &200_u32, &xlm_token);
-        client
-    }
-
-    fn default_params(env: &Env) -> CreateMarketParams {
-        let now = env.ledger().timestamp();
-        CreateMarketParams {
-            title: String::from_str(env, "Will it rain?"),
-            description: String::from_str(env, "Daily weather market"),
-            category: Symbol::new(env, "Sports"),
-            outcomes: vec![env, symbol_short!("yes"), symbol_short!("no")],
-            end_time: now + 1000,
-            resolution_time: now + 2000,
-            dispute_window: 86_400,
-            creator_fee_bps: 100,
-            min_stake: 10_000_000,
-            max_stake: 100_000_000,
-            is_public: true,
-        }
-    }
-
-    #[test]
-    fn create_market_success_returns_incremented_id() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-        let creator = Address::generate(&env);
-
-        let id = client.create_market(&creator, &default_params(&env));
-        assert_eq!(id, 1);
-
-        let id2 = client.create_market(&creator, &default_params(&env));
-        assert_eq!(id2, 2);
-    }
-
-    #[test]
-    fn create_market_fails_end_time_in_past() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-        let creator = Address::generate(&env);
-
-        let mut p = default_params(&env);
-        p.end_time = env.ledger().timestamp(); // not strictly after now
-
-        let result = client.try_create_market(&creator, &p);
-        assert!(matches!(
-            result,
-            Err(Ok(InsightArenaError::InvalidTimeRange))
-        ));
-    }
-
-    #[test]
-    fn create_market_fails_resolution_before_end() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-        let creator = Address::generate(&env);
-
-        let mut p = default_params(&env);
-        p.resolution_time = p.end_time - 1;
-
-        let result = client.try_create_market(&creator, &p);
-        assert!(matches!(
-            result,
-            Err(Ok(InsightArenaError::InvalidTimeRange))
-        ));
-    }
-
-    #[test]
-    fn create_market_fails_single_outcome() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-        let creator = Address::generate(&env);
-
-        let mut p = default_params(&env);
-        p.outcomes = vec![&env, symbol_short!("yes")];
-
-        let result = client.try_create_market(&creator, &p);
-        assert!(matches!(result, Err(Ok(InsightArenaError::InvalidInput))));
-    }
-
-    #[test]
-    fn create_market_fails_fee_too_high() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-        let creator = Address::generate(&env);
-
-        let mut p = default_params(&env);
-        p.creator_fee_bps = 501; // exceeds 500 bps cap
-
-        let result = client.try_create_market(&creator, &p);
-        assert!(matches!(result, Err(Ok(InsightArenaError::InvalidFee))));
-    }
-
-    #[test]
-    fn create_market_fails_when_paused() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-        let creator = Address::generate(&env);
-
-        client.set_paused(&true);
-        let result = client.try_create_market(&creator, &default_params(&env));
-        assert!(matches!(result, Err(Ok(InsightArenaError::Paused))));
-    }
-
-    #[test]
-    fn create_market_fails_stake_too_low() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-        let creator = Address::generate(&env);
-
-        let mut p = default_params(&env);
-        p.min_stake = 1; // below 10_000_000 stroops platform floor
-
-        let result = client.try_create_market(&creator, &p);
-        assert!(matches!(result, Err(Ok(InsightArenaError::StakeTooLow))));
-    }
-
-    #[test]
-    fn create_market_fails_when_category_not_whitelisted() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-        let creator = Address::generate(&env);
-
-        let mut p = default_params(&env);
-        p.category = Symbol::new(&env, "Weather");
-
-        let result = client.try_create_market(&creator, &p);
-        assert!(matches!(result, Err(Ok(InsightArenaError::InvalidInput))));
-    }
-
-    #[test]
-    fn list_categories_returns_seeded_defaults() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-        let categories = client.list_categories();
-
-        assert!(categories.contains(Symbol::new(&env, "Sports")));
-        assert!(categories.contains(Symbol::new(&env, "Crypto")));
-        assert!(categories.contains(Symbol::new(&env, "Politics")));
-        assert!(categories.contains(Symbol::new(&env, "Entertainment")));
-        assert!(categories.contains(Symbol::new(&env, "Science")));
-        assert!(categories.contains(Symbol::new(&env, "Other")));
-    }
-
-    #[test]
-    fn add_category_allows_admin_to_extend_whitelist() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, admin, _oracle) = deploy_with_actors(&env);
-        let weather = Symbol::new(&env, "Weather");
-
-        client.add_category(&admin, &weather);
-
-        let categories = client.list_categories();
-        assert!(categories.contains(weather));
-    }
-
-    #[test]
-    fn remove_category_blocks_future_market_creation() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, admin, _oracle) = deploy_with_actors(&env);
-        let creator = Address::generate(&env);
-        let science = Symbol::new(&env, "Science");
-
-        client.remove_category(&admin, &science);
-
-        let categories = client.list_categories();
-        assert!(!categories.contains(science.clone()));
-
-        let mut p = default_params(&env);
-        p.category = science;
-        let result = client.try_create_market(&creator, &p);
-        assert!(matches!(result, Err(Ok(InsightArenaError::InvalidInput))));
-    }
-
-    #[test]
-    fn non_admin_cannot_mutate_categories() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, _admin, _oracle) = deploy_with_actors(&env);
-        let random = Address::generate(&env);
-        let weather = Symbol::new(&env, "Weather");
-        let sports = Symbol::new(&env, "Sports");
-
-        let add_result = client.try_add_category(&random, &weather);
-        assert!(matches!(
-            add_result,
-            Err(Ok(InsightArenaError::Unauthorized))
-        ));
-
-        let remove_result = client.try_remove_category(&random, &sports);
-        assert!(matches!(
-            remove_result,
-            Err(Ok(InsightArenaError::Unauthorized))
-        ));
-    }
-
-    // ── get_market ────────────────────────────────────────────────────────────
-
-    #[test]
-    fn get_market_returns_correct_market() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-        let creator = Address::generate(&env);
-
-        let id = client.create_market(&creator, &default_params(&env));
-        let market = client.get_market(&id);
-        assert_eq!(market.market_id, id);
-        assert_eq!(market.creator, creator);
-    }
-
-    #[test]
-    fn get_market_returns_not_found_for_missing_id() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-
-        let result = client.try_get_market(&99_u64);
-        assert!(matches!(result, Err(Ok(InsightArenaError::MarketNotFound))));
-    }
-
-    // ── get_market_count ──────────────────────────────────────────────────────
-
-    #[test]
-    fn get_market_count_zero_before_any_market() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-
-        assert_eq!(client.get_market_count(), 0);
-    }
-
-    #[test]
-    fn get_market_count_increments_with_each_market() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-        let creator = Address::generate(&env);
-
-        client.create_market(&creator, &default_params(&env));
-        assert_eq!(client.get_market_count(), 1);
-
-        client.create_market(&creator, &default_params(&env));
-        assert_eq!(client.get_market_count(), 2);
-    }
-
-    // ── list_markets ──────────────────────────────────────────────────────────
-
-    #[test]
-    fn list_markets_empty_when_no_markets() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-
-        let list = client.list_markets(&1_u64, &10_u32);
-        assert_eq!(list.len(), 0);
-    }
-
-    #[test]
-    fn get_markets_by_category_returns_paginated_results() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-        let creator = Address::generate(&env);
-
-        let sports_category = Symbol::new(&env, "Sports");
-        let crypto_category = Symbol::new(&env, "Crypto");
-
-        let first_sports = client.create_market(&creator, &default_params(&env));
-
-        let mut crypto = default_params(&env);
-        crypto.category = crypto_category;
-        client.create_market(&creator, &crypto);
-
-        let second_sports_id = client.create_market(&creator, &default_params(&env));
-        let third_sports_id = client.create_market(&creator, &default_params(&env));
-
-        let first_page = client.get_markets_by_category(&sports_category, &0_u64, &2_u32);
-        assert_eq!(first_page.len(), 2);
-        assert_eq!(first_page.get(0).unwrap().market_id, first_sports);
-        assert_eq!(first_page.get(1).unwrap().market_id, second_sports_id);
-
-        let second_page = client.get_markets_by_category(&sports_category, &2_u64, &2_u32);
-        assert_eq!(second_page.len(), 1);
-        assert_eq!(second_page.get(0).unwrap().market_id, third_sports_id);
-    }
-
-    #[test]
-    fn category_index_is_kept_in_sync_on_market_creation() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-        let creator = Address::generate(&env);
-        let sports = Symbol::new(&env, "Sports");
-
-        let first_id = client.create_market(&creator, &default_params(&env));
-
-        let mut crypto = default_params(&env);
-        crypto.category = Symbol::new(&env, "Crypto");
-        client.create_market(&creator, &crypto);
-
-        let second_id = client.create_market(&creator, &default_params(&env));
-
-        let stored_index = env.as_contract(&client.address, || {
-            env.storage()
-                .persistent()
-                .get::<DataKey, Vec<u64>>(&DataKey::CategoryIndex(sports.clone()))
-                .unwrap()
-        });
-
-        assert_eq!(stored_index.len(), 2);
-        assert_eq!(stored_index.get(0), Some(first_id));
-        assert_eq!(stored_index.get(1), Some(second_id));
-    }
-
-    #[test]
-    fn list_markets_returns_all_when_within_limit() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-        let creator = Address::generate(&env);
-
-        for _ in 0..3 {
-            client.create_market(&creator, &default_params(&env));
-        }
-
-        let list = client.list_markets(&1_u64, &10_u32);
-        assert_eq!(list.len(), 3);
-        assert_eq!(list.get(0).unwrap().market_id, 1);
-        assert_eq!(list.get(2).unwrap().market_id, 3);
-    }
-
-    #[test]
-    fn list_markets_respects_pagination_start() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-        let creator = Address::generate(&env);
-
-        for _ in 0..5 {
-            client.create_market(&creator, &default_params(&env));
-        }
-
-        // Start from market ID 3, take up to 10
-        let list = client.list_markets(&3_u64, &10_u32);
-        assert_eq!(list.len(), 3); // IDs 3, 4, 5
-        assert_eq!(list.get(0).unwrap().market_id, 3);
-    }
-
-    #[test]
-    fn list_markets_caps_at_max_limit_50() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-        let creator = Address::generate(&env);
-
-        for _ in 0..60 {
-            client.create_market(&creator, &default_params(&env));
-        }
-
-        let list = client.list_markets(&1_u64, &100_u32); // ask for 100, should get 50
-        assert_eq!(list.len(), 50);
-    }
-
-    #[test]
-    fn list_markets_empty_when_start_out_of_bounds() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let client = deploy(&env);
-        let creator = Address::generate(&env);
-
-        client.create_market(&creator, &default_params(&env));
-
-        // start > total count → empty
-        let list = client.list_markets(&99_u64, &10_u32);
-        assert_eq!(list.len(), 0);
-    }
-
-    // ── close_market ──────────────────────────────────────────────────────────
-
-    /// Helper: deploy a contract and return client together with pre-registered
-    /// admin and oracle addresses (the same ones used during `initialize`).
-    fn deploy_with_actors(env: &Env) -> (InsightArenaContractClient<'_>, Address, Address) {
-        let id = env.register(InsightArenaContract, ());
-        let client = InsightArenaContractClient::new(env, &id);
-        let admin = Address::generate(env);
-        let oracle = Address::generate(env);
-        let xlm_token = register_token(env);
-        env.mock_all_auths();
-        client.initialize(&admin, &oracle, &200_u32, &xlm_token);
-        (client, admin, oracle)
-    }
-
-    /// Helper: deploy contract, return client + admin + oracle + token address.
-    fn deploy_with_token(env: &Env) -> (InsightArenaContractClient<'_>, Address, Address, Address) {
-        let id = env.register(InsightArenaContract, ());
-        let client = InsightArenaContractClient::new(env, &id);
-        let admin = Address::generate(env);
-        let oracle = Address::generate(env);
-        let xlm_token = register_token(env);
-        env.mock_all_auths();
-        client.initialize(&admin, &oracle, &200_u32, &xlm_token);
-        (client, admin, oracle, xlm_token)
-    }
-
-    // (a) close_market called before end_time → MarketStillOpen
-    #[test]
-    fn close_market_fails_before_end_time() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, _admin, oracle) = deploy_with_actors(&env);
-        let creator = Address::generate(&env);
-
-        // Market end_time is now + 1000; current timestamp is still "now"
-        let id = client.create_market(&creator, &default_params(&env));
-
-        let result = client.try_close_market(&oracle, &id);
-        assert!(matches!(
-            result,
-            Err(Ok(InsightArenaError::MarketStillOpen))
-        ));
-    }
-
-    // (b) close_market called after end_time by the oracle → success + is_closed == true
-    #[test]
-    fn close_market_success_by_oracle_after_end_time() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, _admin, oracle) = deploy_with_actors(&env);
-        let creator = Address::generate(&env);
-
-        let id = client.create_market(&creator, &default_params(&env));
-
-        // Advance ledger time past end_time (now + 1000)
-        env.ledger().set_timestamp(env.ledger().timestamp() + 1001);
-
-        client.close_market(&oracle, &id);
-
-        let market = client.get_market(&id);
-        assert!(market.is_closed);
-        assert!(!market.is_resolved);
-    }
-
-    // (b-alt) close_market called after end_time by the admin → success
-    #[test]
-    fn close_market_success_by_admin_after_end_time() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, admin, _oracle) = deploy_with_actors(&env);
-        let creator = Address::generate(&env);
-
-        let id = client.create_market(&creator, &default_params(&env));
-
-        env.ledger().set_timestamp(env.ledger().timestamp() + 1001);
-
-        client.close_market(&admin, &id);
-
-        let market = client.get_market(&id);
-        assert!(market.is_closed);
-    }
-
-    // (c) double-close attempt → MarketAlreadyResolved not triggered, but a
-    //     second close on an already-closed (not yet resolved) market succeeds
-    //     because is_resolved is still false; however once resolved it must fail.
-    //     We test the resolved path: set is_resolved manually via a resolved market
-    //     scenario by directly checking that a market flagged resolved returns the error.
-    //
-    //     Since we can only interact through the public ABI, we test the reachable
-    //     path: close a market that has already been resolved (simulated by calling
-    //     close twice — second call must still pass because is_resolved stays false
-    //     until resolve_market is implemented).  Instead we verify that calling
-    //     close on a non-existent market returns MarketNotFound, and that calling
-    //     close on an already-closed-then-externally-resolved market returns
-    //     MarketAlreadyResolved via direct storage manipulation in the test.
-    #[test]
-    fn close_market_fails_when_already_resolved() {
-        use crate::storage_types::{DataKey, Market};
-
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, _admin, oracle) = deploy_with_actors(&env);
-        let creator = Address::generate(&env);
-
-        let id = client.create_market(&creator, &default_params(&env));
-
-        // Advance past end_time and close the market normally
-        env.ledger().set_timestamp(env.ledger().timestamp() + 1001);
-        client.close_market(&oracle, &id);
-
-        // Simulate resolution by mutating the stored market directly using the
-        // correct contract address from the deployed client.
-        let contract_id = client.address.clone();
-        let mut market: Market = env.as_contract(&contract_id, || {
-            env.storage()
-                .persistent()
-                .get(&DataKey::Market(id))
-                .unwrap()
-        });
-        market.is_resolved = true;
-        env.as_contract(&contract_id, || {
-            env.storage()
-                .persistent()
-                .set(&DataKey::Market(id), &market);
-        });
-
-        // Now try to close again — should fail with MarketAlreadyResolved
-        let result = client.try_close_market(&oracle, &id);
-        assert!(matches!(
-            result,
-            Err(Ok(InsightArenaError::MarketAlreadyResolved))
-        ));
-    }
-
-    // ── cancel_market ─────────────────────────────────────────────────────────
-
-    // (a) Non-admin caller → Unauthorized
-    #[test]
-    fn cancel_market_fails_for_non_admin() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, _admin, _oracle, _token) = deploy_with_token(&env);
-        let creator = Address::generate(&env);
-        let random = Address::generate(&env);
-
-        let id = client.create_market(&creator, &default_params(&env));
-
-        let result = client.try_cancel_market(&random, &id);
-        assert!(matches!(result, Err(Ok(InsightArenaError::Unauthorized))));
-    }
-
-    // (b) Unknown market_id → MarketNotFound
-    #[test]
-    fn cancel_market_fails_market_not_found() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, admin, _oracle, _token) = deploy_with_token(&env);
-
-        let result = client.try_cancel_market(&admin, &99_u64);
-        assert!(matches!(result, Err(Ok(InsightArenaError::MarketNotFound))));
-    }
-
-    // (c) Already-resolved market → MarketAlreadyResolved
-    #[test]
-    fn cancel_market_fails_when_already_resolved() {
-        use crate::storage_types::{DataKey, Market};
-
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, admin, _oracle, _token) = deploy_with_token(&env);
-        let creator = Address::generate(&env);
-
-        let id = client.create_market(&creator, &default_params(&env));
-
-        // Simulate resolution via direct storage mutation.
-        let contract_id = client.address.clone();
-        let mut market: Market = env.as_contract(&contract_id, || {
-            env.storage()
-                .persistent()
-                .get(&DataKey::Market(id))
-                .unwrap()
-        });
-        market.is_resolved = true;
-        env.as_contract(&contract_id, || {
-            env.storage()
-                .persistent()
-                .set(&DataKey::Market(id), &market);
-        });
-
-        let result = client.try_cancel_market(&admin, &id);
-        assert!(matches!(
-            result,
-            Err(Ok(InsightArenaError::MarketAlreadyResolved))
-        ));
-    }
-
-    // (d) Double-cancel → MarketAlreadyCancelled
-    #[test]
-    fn cancel_market_fails_when_already_cancelled() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, admin, _oracle, _token) = deploy_with_token(&env);
-        let creator = Address::generate(&env);
-
-        let id = client.create_market(&creator, &default_params(&env));
-        client.cancel_market(&admin, &id); // first cancel succeeds
-
-        let result = client.try_cancel_market(&admin, &id);
-        assert!(matches!(
-            result,
-            Err(Ok(InsightArenaError::MarketAlreadyCancelled))
-        ));
-    }
-
-    // (e) Successful cancel with no predictors → market.is_cancelled == true,
-    //     MarketCancelled event emitted, no refund calls made.
-    #[test]
-    fn cancel_market_success_no_predictors() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, admin, _oracle, _token) = deploy_with_token(&env);
-        let creator = Address::generate(&env);
-
-        let id = client.create_market(&creator, &default_params(&env));
-        client.cancel_market(&admin, &id);
-
-        let market = client.get_market(&id);
-        assert!(market.is_cancelled);
-        assert!(!market.is_resolved);
-    }
-
-    // (f) Cancel with multiple predictors → all stakes refunded, balances restored.
-    //
-    // Because no `predict` function exists yet, predictions are seeded directly
-    // into persistent storage (same technique as the close_market resolved test).
-    // The contract escrow balance is pre-funded by minting tokens to the contract.
-    #[test]
-    fn cancel_market_refunds_all_predictors() {
-        use crate::storage_types::{DataKey, Prediction};
-        use soroban_sdk::token::{Client as TokenClient, StellarAssetClient};
-
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, admin, _oracle, xlm_token) = deploy_with_token(&env);
-        let creator = Address::generate(&env);
-
-        let id = client.create_market(&creator, &default_params(&env));
-
-        // Prepare two predictors with distinct stakes.
-        let predictor_a = Address::generate(&env);
-        let predictor_b = Address::generate(&env);
-        let stake_a: i128 = 20_000_000; // 2 XLM
-        let stake_b: i128 = 50_000_000; // 5 XLM
-
-        let contract_id = client.address.clone();
-
-        // Seed Prediction records and PredictorList directly into contract storage.
-        env.as_contract(&contract_id, || {
-            let pred_a = Prediction::new(
-                id,
-                predictor_a.clone(),
-                symbol_short!("yes"),
-                stake_a,
-                env.ledger().timestamp(),
-            );
-            let pred_b = Prediction::new(
-                id,
-                predictor_b.clone(),
-                symbol_short!("no"),
-                stake_b,
-                env.ledger().timestamp(),
-            );
-
-            env.storage()
-                .persistent()
-                .set(&DataKey::Prediction(id, predictor_a.clone()), &pred_a);
-            env.storage()
-                .persistent()
-                .set(&DataKey::Prediction(id, predictor_b.clone()), &pred_b);
-
-            let mut predictors = soroban_sdk::Vec::new(&env);
-            predictors.push_back(predictor_a.clone());
-            predictors.push_back(predictor_b.clone());
-            env.storage()
-                .persistent()
-                .set(&DataKey::PredictorList(id), &predictors);
-        });
-
-        // Fund the contract escrow with the total staked amount.
-        let total_staked = stake_a + stake_b;
-        StellarAssetClient::new(&env, &xlm_token).mint(&contract_id, &total_staked);
-
-        // Confirm predictors start with zero balance.
-        let token_client = TokenClient::new(&env, &xlm_token);
-        assert_eq!(token_client.balance(&predictor_a), 0);
-        assert_eq!(token_client.balance(&predictor_b), 0);
-
-        // Cancel the market.
-        client.cancel_market(&admin, &id);
-
-        // Every predictor must receive exactly their stake back.
-        assert_eq!(token_client.balance(&predictor_a), stake_a);
-        assert_eq!(token_client.balance(&predictor_b), stake_b);
-
-        // Market must be flagged as cancelled.
-        let market = client.get_market(&id);
-        assert!(market.is_cancelled);
-    }
 }
